test: reset properties after E2Es and move them to auto creation suite
