+++ test-clients/src/main/java/com/hedera/services/bdd/suites/crypto/AutoAccountCreationSuite.java
@@ -15,6 +15,7 @@
  */
 package com.hedera.services.bdd.suites.crypto;
 
+import static com.google.protobuf.ByteString.copyFromUtf8;
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
 import static com.hedera.services.bdd.spec.HapiPropertySource.asSolidityAddress;
 import static com.hedera.services.bdd.spec.PropertySource.asAccount;
@@ -60,7 +61,12 @@ import static com.hedera.services.bdd.spec.utilops.UtilVerbs.resetToDefault;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sourcing;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
 import static com.hedera.services.bdd.suites.contract.Utils.aaWith;
+import static com.hedera.services.bdd.suites.contract.Utils.accountId;
+import static com.hedera.services.bdd.suites.contract.Utils.ocWith;
 import static com.hedera.services.bdd.suites.contract.hapi.ContractUpdateSuite.ADMIN_KEY;
+import static com.hedera.services.bdd.suites.crypto.CryptoCreateSuite.LAZY_CREATION_ENABLED;
+import static com.hedera.services.bdd.suites.crypto.CryptoCreateSuite.TRUE;
+import static com.hedera.services.ethereum.EthTxSigs.recoverAddressFromPubKey;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.ACCOUNT_DELETED;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.ACCOUNT_REPEATED_IN_ACCOUNT_AMOUNTS;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_ACCOUNT_BALANCE;
@@ -90,6 +96,7 @@ import com.hederahashgraph.api.proto.java.TokenSupplyType;
 import com.hederahashgraph.api.proto.java.TokenTransferList;
 import com.hederahashgraph.api.proto.java.TokenType;
 import com.hederahashgraph.api.proto.java.TransactionRecord;
+import com.hederahashgraph.api.proto.java.TransferList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
@@ -141,6 +148,9 @@ public class AutoAccountCreationSuite extends HapiApiSuite {
     private static final long EXPECTED_HBAR_TRANSFER_AUTO_CREATION_FEE = 39418863L;
     private static final long EXPECTED_MULTI_TOKEN_TRANSFER_AUTO_CREATION_FEE = 42427268L;
     private static final long EXPECTED_SINGLE_TOKEN_TRANSFER_AUTO_CREATE_FEE = 40927290L;
+    private static final String HBAR_XFER = "hbarXfer";
+    private static final String NFT_XFER = "nftXfer";
+    private static final String FT_XFER = "ftXfer";
 
     public static void main(String... args) {
         new AutoAccountCreationSuite().runSuiteSync();
@@ -182,7 +192,10 @@ public class AutoAccountCreationSuite extends HapiApiSuite {
                 autoCreateWithNftFallBackFeeFails(),
                 repeatedAliasInSameTransferListFails(),
                 tokenTransfersFailWhenFeatureFlagDisabled(),
-                canAutoCreateWithHbarAndTokenTransfers());
+                canAutoCreateWithHbarAndTokenTransfers(),
+                transferHbarsToEVMAddressAlias(),
+                transferFungibleToEVMAddressAlias(),
+                transferNonFungibleToEVMAddressAlias());
     }
 
     private HapiApiSpec canAutoCreateWithHbarAndTokenTransfers() {
@@ -223,7 +236,6 @@ public class AutoAccountCreationSuite extends HapiApiSuite {
     private HapiApiSpec tokenTransfersFailWhenFeatureFlagDisabled() {
         final var initialTokenSupply = 1000;
         final var fungibleTokenXfer = "fungibleTokenXfer";
-        final var nftXfer = "nftXfer";
 
         return defaultHapiSpec("tokenTransfersFailWhenFeatureFlagDisabled")
                 .given(
@@ -271,12 +283,14 @@ public class AutoAccountCreationSuite extends HapiApiSuite {
                         cryptoTransfer(
                                         movingUnique(NFT_INFINITE_SUPPLY_TOKEN, 1, 2)
                                                 .between(CIVILIAN, VALID_ALIAS))
-                                .via(nftXfer)
+                                .via(NFT_XFER)
                                 .payingWith(CIVILIAN)
                                 .hasKnownStatus(NOT_SUPPORTED)
                                 .logged(),
                         getTxnRecord(fungibleTokenXfer).andAllChildRecords().hasChildRecordCount(0),
-                        getTxnRecord(nftXfer).andAllChildRecords().hasNonStakingChildRecordCount(0),
+                        getTxnRecord(NFT_XFER)
+                                .andAllChildRecords()
+                                .hasNonStakingChildRecordCount(0),
                         /* --- hbar auto creations should still pass */
                         cryptoTransfer(tinyBarsFromToWithAlias(CIVILIAN, VALID_ALIAS, ONE_HBAR))
                                 .payingWith(CIVILIAN)
@@ -1397,5 +1411,162 @@ public class AutoAccountCreationSuite extends HapiApiSuite {
                                                 .balance(
                                                         (INITIAL_BALANCE * ONE_HBAR)
                                                                 - 3 * ONE_HUNDRED_HBARS)));
+    }
+
+    private HapiApiSpec transferHbarsToEVMAddressAlias() {
+
+        final AtomicReference<AccountID> partyId = new AtomicReference<>();
+        final AtomicReference<ByteString> partyAlias = new AtomicReference<>();
+        final AtomicReference<ByteString> counterAlias = new AtomicReference<>();
+
+        return defaultHapiSpec("TransferHbarsToEVMAddressAlias")
+                .given(
+                        overriding(LAZY_CREATION_ENABLED, TRUE),
+                        cryptoCreate(PARTY).maxAutomaticTokenAssociations(2),
+                        newKeyNamed(SECP_256K1_SOURCE_KEY).shape(SECP_256K1_SHAPE),
+                        withOpContext(
+                                (spec, opLog) -> {
+                                    final var registry = spec.registry();
+                                    final var ecdsaKey = registry.getKey(SECP_256K1_SOURCE_KEY);
+                                    final var tmp = ecdsaKey.getECDSASecp256K1().toByteArray();
+                                    final var addressBytes = recoverAddressFromPubKey(tmp);
+                                    assert addressBytes != null;
+                                    final var evmAddressBytes = ByteString.copyFrom(addressBytes);
+                                    partyId.set(registry.getAccountID(PARTY));
+                                    partyAlias.set(
+                                            ByteString.copyFrom(asSolidityAddress(partyId.get())));
+                                    counterAlias.set(evmAddressBytes);
+                                }))
+                .when(
+                        cryptoTransfer(
+                                        (spec, b) ->
+                                                b.setTransfers(
+                                                        TransferList.newBuilder()
+                                                                .addAccountAmounts(
+                                                                        aaWith(
+                                                                                partyAlias.get(),
+                                                                                -2))
+                                                                .addAccountAmounts(
+                                                                        aaWith(
+                                                                                counterAlias.get(),
+                                                                                +2))))
+                                .signedBy(DEFAULT_PAYER, PARTY)
+                                .via(HBAR_XFER))
+                .then(
+                        getTxnRecord(HBAR_XFER)
+                                .hasChildRecordCount(1)
+                                .hasChildRecords(recordWith().status(SUCCESS).memo(LAZY_MEMO)),
+                        resetToDefault(LAZY_CREATION_ENABLED));
+    }
+
+    private HapiApiSpec transferFungibleToEVMAddressAlias() {
+
+        final var fungibleToken = "fungibleToken";
+        final AtomicReference<TokenID> ftId = new AtomicReference<>();
+        final AtomicReference<AccountID> partyId = new AtomicReference<>();
+        final AtomicReference<ByteString> partyAlias = new AtomicReference<>();
+        final AtomicReference<ByteString> counterAlias = new AtomicReference<>();
+
+        return defaultHapiSpec("TransferFungibleToEVMAddressAlias")
+                .given(
+                        overriding(LAZY_CREATION_ENABLED, TRUE),
+                        cryptoCreate(PARTY).maxAutomaticTokenAssociations(2),
+                        tokenCreate(fungibleToken).treasury(PARTY).initialSupply(1_000_000),
+                        newKeyNamed(SECP_256K1_SOURCE_KEY).shape(SECP_256K1_SHAPE),
+                        withOpContext(
+                                (spec, opLog) -> {
+                                    final var registry = spec.registry();
+                                    final var ecdsaKey = registry.getKey(SECP_256K1_SOURCE_KEY);
+                                    final var tmp = ecdsaKey.getECDSASecp256K1().toByteArray();
+                                    final var addressBytes = recoverAddressFromPubKey(tmp);
+                                    assert addressBytes != null;
+                                    final var evmAddressBytes = ByteString.copyFrom(addressBytes);
+                                    ftId.set(registry.getTokenID(fungibleToken));
+                                    partyId.set(registry.getAccountID(PARTY));
+                                    partyAlias.set(
+                                            ByteString.copyFrom(asSolidityAddress(partyId.get())));
+                                    counterAlias.set(evmAddressBytes);
+                                }))
+                .when(
+                        cryptoTransfer(
+                                        (spec, b) ->
+                                                b.addTokenTransfers(
+                                                        TokenTransferList.newBuilder()
+                                                                .setToken(ftId.get())
+                                                                .addTransfers(
+                                                                        aaWith(
+                                                                                partyAlias.get(),
+                                                                                -500))
+                                                                .addTransfers(
+                                                                        aaWith(
+                                                                                counterAlias.get(),
+                                                                                +500))))
+                                .signedBy(DEFAULT_PAYER, PARTY)
+                                .via(FT_XFER))
+                .then(
+                        getTxnRecord(FT_XFER)
+                                .hasChildRecordCount(1)
+                                .hasChildRecords(recordWith().status(SUCCESS).memo(LAZY_MEMO)),
+                        resetToDefault(LAZY_CREATION_ENABLED));
+    }
+
+    private HapiApiSpec transferNonFungibleToEVMAddressAlias() {
+
+        final var nonFungibleToken = "nonFungibleToken";
+        final AtomicReference<TokenID> nftId = new AtomicReference<>();
+        final AtomicReference<AccountID> partyId = new AtomicReference<>();
+        final AtomicReference<ByteString> partyAlias = new AtomicReference<>();
+        final AtomicReference<ByteString> counterAlias = new AtomicReference<>();
+
+        return defaultHapiSpec("TransferNonFungibleToEVMAddressAlias")
+                .given(
+                        overriding(LAZY_CREATION_ENABLED, TRUE),
+                        newKeyNamed(MULTI_KEY),
+                        newKeyNamed(SECP_256K1_SOURCE_KEY).shape(SECP_256K1_SHAPE),
+                        cryptoCreate(PARTY).maxAutomaticTokenAssociations(2),
+                        tokenCreate(nonFungibleToken)
+                                .initialSupply(0)
+                                .treasury(PARTY)
+                                .tokenType(NON_FUNGIBLE_UNIQUE)
+                                .supplyKey(MULTI_KEY),
+                        mintToken(
+                                nonFungibleToken,
+                                List.of(copyFromUtf8("Test transfer nft to EVM address alias."))),
+                        withOpContext(
+                                (spec, opLog) -> {
+                                    final var registry = spec.registry();
+                                    final var ecdsaKey = registry.getKey(SECP_256K1_SOURCE_KEY);
+                                    final var tmp = ecdsaKey.getECDSASecp256K1().toByteArray();
+                                    final var addressBytes = recoverAddressFromPubKey(tmp);
+                                    assert addressBytes != null;
+                                    final var evmAddressBytes = ByteString.copyFrom(addressBytes);
+                                    nftId.set(registry.getTokenID(nonFungibleToken));
+                                    partyId.set(registry.getAccountID(PARTY));
+                                    partyAlias.set(
+                                            ByteString.copyFrom(asSolidityAddress(partyId.get())));
+                                    counterAlias.set(evmAddressBytes);
+                                }))
+                .when(
+                        cryptoTransfer(
+                                        (spec, b) ->
+                                                b.addTokenTransfers(
+                                                        TokenTransferList.newBuilder()
+                                                                .setToken(nftId.get())
+                                                                .addNftTransfers(
+                                                                        ocWith(
+                                                                                accountId(
+                                                                                        partyAlias
+                                                                                                .get()),
+                                                                                accountId(
+                                                                                        counterAlias
+                                                                                                .get()),
+                                                                                1L))))
+                                .signedBy(DEFAULT_PAYER, PARTY)
+                                .via(NFT_XFER))
+                .then(
+                        getTxnRecord(NFT_XFER)
+                                .hasChildRecordCount(1)
+                                .hasChildRecords(recordWith().status(SUCCESS).memo(LAZY_MEMO)),
+                        resetToDefault(LAZY_CREATION_ENABLED));
     }
 }
